chore(mechanics): code cleanup
